requirements.txt
anthropic>=0.84.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
flask>=3.0.0
gunicorn>=22.0.0
scikit-learn>=1.3.0
azure-storage-blob>=12.13.0
azure-identity>=1.13.0
# sentence-transformers is an optional upgrade for the semantic retrieval branch.
# Install it to use BERT-based embeddings instead of the default LSA fallback:
#   pip install sentence-transformers>=2.2.0
